Change how we check whether output_dir is set
The prior approach is the better one, but I just can’t get it to work,
and I don’t understand why. (It always reports that the variable isn’t
set, though it certainly is.) So I’m using the lazy approach, which
doesn’t distinguish between unset and null. But I suppose this approach
is better, because unset, null…either way it’s not useful here.