Change how we handle shutdowns
There's some slightly convoluted logic here around shutting down, because otherwise it would be impossible to start the machine and do anything with it unless there were videos in the queue. (That is, it would shut down immediately for lack of videos.) The solution here is to pause for 3 minutes (180 seconds), then check if any users are logged in. If not, it shuts down.