Replace M3U functionality with something simpler
Instead of dealing with the M3U file and using ffmpeg's ability to go through that and fetch files, instead we just incrementally fetch segments until we get an error, and then we use ffmpeg's concat functionality.

The only reason for this change is that ffmpeg -protocol_whitelist isn't working. ffmpeg says it's invalid. It's not! It's very much a function of ffmpeg, and has been for a long time! But after 30 minutes down that rabbit hole, I figured I'd take a different approach.